changes made to models/users and does not work
