+++ server/models/users.js
@@ -1,4 +1,8 @@
 var db = require('../database');
+const bcrypt = require('bcryptjs');
+const passport = require('passport');
+const saltRounds = 4;
+
 var users = {
   get: function(callback) {
     return db.query('select * from users', callback);
@@ -8,9 +12,28 @@ var users = {
   },
   add: function( users , callback) {
     return db.query(
-      'insert into users values(?,?,?,?)',
-      [users.id, users.name, users.email, users.password],
-      callback
+      passport.authenticate('basic', { session: false }),
+      (req, res) => {
+      let email = req.body.email.trim();
+      let password = req.body.password.trim();
+    
+      if((typeof email === "string") &&
+         (typeof password === "string"))
+      {
+        bcrypt.hash(password, saltRounds).then(hash =>
+          db.query('INSERT INTO users (email, password) VALUES (?,?)', [email, hash])
+        )
+        .then(dbResults => {
+            console.log(dbResults);
+            res.sendStatus(201);
+        })
+        .catch(error => res.sendStatus(500));
+      }
+      else {
+        console.log("incorrect email or password, both must be strings");
+        res.sendStatus(400);
+      }
+    }
     );
   },
   delete: function(id, callback) {
