+++ server/routes/users.js
@@ -1,11 +1,9 @@
 var express = require('express');
 var router = express.Router();
 var users = require('../models/users');
+const Strategy = require('passport-http').BasicStrategy;
 const bcrypt = require('bcryptjs');
 const passport = require('passport');
-const Strategy = require('passport-http').BasicStrategy;
-
-const saltrounds = 4;
 
 passport.use(new Strategy((email, password, cb) => {
   db.query('SELECT id, email, password FROM users WHERE email = ?', [email]).then(dbResults => {
@@ -59,29 +57,6 @@ router.post('/', function(req, res, next) {
   });
 });
 
-  {/*router.post('/', 
-  passport.authenticate('basic', { session: false }),
-  (req, res) => {
-  let email = req.body.email.trim();
-  let password = req.body.password.trim();
-
-  if((typeof email === "string") &&
-     (typeof password === "string"))
-  {
-    bcrypt.hash(password, saltRounds).then(hash =>
-      db.query('INSERT INTO users (email, password) VALUES (?,?)', [email, hash])
-    )
-    .then(dbResults => {
-        console.log(dbResults);
-        res.sendStatus(201);
-    })
-    .catch(error => res.sendStatus(500));
-  }
-  else {
-    console.log("incorrect email or password, both must be strings");
-    res.sendStatus(400);
-  }
-})*/}
 
 router.delete('/:id', 
   passport.authenticate('basic', { session: false }),
